requirements.txt
Django==1.6.5
Fabric==1.10.0
South==1.0
argparse==1.2.1
distribute==0.6.24
django-filter==0.7
djangorestframework==2.3.14
ecdsa==0.11
paramiko==1.14.1
pycrypto==2.6.1
pytz==2014.7
wsgiref==0.1.2
